+++ apps/pika-chat/src/lib/client/features/site-admin/pages/general-settings.svelte
@@ -22,9 +22,6 @@
 
     let { pageHeaderRight = $bindable() }: Props = $props();
 
-    //TODO: fix any
-    pageHeaderRight = pageHeaderRightSnippet as any;
-
     const features = $derived(siteAdmin.siteFeatures);
 
     function formatUserTypes(userTypes?: string[]): string {