+++ apps/pika-chat/src/lib/client/features/site-admin/pages/instruction-augmentation.svelte
@@ -30,8 +30,7 @@
 
     $effect(() => {
         setTimeout(() => {
-            //TODO: fix any
-            pageHeaderRight = pageHeaderRightSnippet as any;
+            pageHeaderRight = pageHeaderRightSnippet;
         }, 1);
     });