+++ apps/pika-chat/src/lib/client/features/site-admin/pages/session-analytics.svelte
@@ -42,8 +42,7 @@
 
     $effect(() => {
         setTimeout(() => {
-            //TODO: fix any
-            pageHeaderRight = pageHeaderRightSnippet as any;
+            pageHeaderRight = pageHeaderRightSnippet;
         }, 1);
     });
 </script>